.history/requirements_20251217055724.txt
# Core AI/ML - FREE LOCAL MODELS
langchain==0.1.0
langchain-community==0.0.13
ollama==0.1.6

# Vector Store & Memory
chromadb==0.4.22
faiss-cpu==1.7.4
sentence-transformers==2.2.2

# WhatsApp Integration
twilio==8.11.1
flask==3.0.0
requests==2.31.0

# Email Integration
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.116.0

# MCP (Model Context Protocol)
mcp==0.1.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Utilities
python-dotenv==1.0.0
redis==5.0.1
sqlalchemy==2.0.25
pymongo==4.6.1

# Security
cryptography==42.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Async & Web
fastapi==0.109.0
uvicorn==0.27.0
aiohttp==3.9.1
httpx==0.26.0

# Logging & Monitoring
loguru==0.7.2
prometheus-client==0.19.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-mock==3.12.0
